Add markdown support for consent message (#2100)
* Add markdown support for consent message

This adds support for rendering consent messages as markdown.

We use [`marked`](https://www.npmjs.com/package/marked) for rendering
the markdown to HTML, and then sanitize the HTML with
[`DOMPurify`](https://www.npmjs.com/package/dompurify). The consent
message is rendered async and the original message (raw markdown) is
shown until the markdown was rendered (in practice, instantaneous).

* Use NNS dapp markdown helper