Pre-fill anchors in VC flow (#2109)
* Pre-fill anchors in VC flow

This updates the storage to store (a digest of) principals of visited
dapps. This allows the VC flow to know (in the general, best case
scenario) which anchor (identity number) the user is intending to use.

The storage migration includes three things:

* Anchors are not written directly as `{ 10000:  <data> }` anymore but
  are nested as `{ anchors: { 10000: <data> } }` to allow for a new
  top-level field `hasher` which is global to all anchors.
* Said new field `hasher` is introduced
* Anchors themselves now have a new field, `knownPrincipalDigests`.

See storage module documentation for more information on the digests.

* Improve principal storage & test

* Prune old principals