Add support for PIN identities in VC
This abstracts the passkey use flow so that it can be used both in the
authenticateBox (where it was already used) and in the VC flow.

The tests are updated so that their number doesn't grow exponentially:

* One test with legacy RP
* One test with legacy ISS
* One test with PIN

By default, the tests run with non-legacy domains and run with webauthn
passkey.