API V2 Refactor
This refactors the existing API v2 to be consistent with new patterns
introduced in the issuer API:
* Use camel case variants
* Use `Result` return values

The `Err` variants will be extended in the future with more types. These
will be breaking changes. However, given we mark the API v2 as experimental
this should be ok.

In order to have CI pass, we have an exception in the form of a patch file
that undoes the breaking changes that we intentionally made. If we introduce
additional breaking changes, the patch will no longer apply and thus make the
check fail.