Introduce API v2 authn_method_metadata_replace
This PR introduces the API v2 `authn_method_metadata_replace` method.
It replaces the metadata map of an existing `authn_method` of an identity.
It will be used in place of the current `update` method. The scope reduction
to just `metadata` is there to have a clear separation between security
relevant changes to an authn_method (such as modifying the `protection`) and
changes not relevant to security (modifying the `alias` entry in the metadata).