Add VC support for alternative origins (#2211)
* Add VC support for alternative origins

This updates the frontend verifiable credentials flow to access a new
`derivationOrigin` field in the VC request.

The frontend then queries the specified derivation origin and checks the
RP is listed as an alternative origin, as in the authentication flow.

If successful, the specified derivation origin is used as the origin for
the VC flow.

The Verifiable Credentials e2e tests are moved to their own directory
and split between base flow tests, utils, and new alternative origin
tests.

* Clean up error message

* Fix up hard coded alternative origins