Clean up compatibility notice (#2215)
* Clean up compatibility notice

This updates the compatibility notice to make it more correct and more
informative.

In particular, it now only ensures that `window.PublicKeyCredential`
exists, which is enough for ensuring webauthn support in most cases. For
the other cases, we simply show an error later on during usage.

The compatibility page is also cleaned up a bit to make sure all the
text ("Your browser doesn't support II...") is inside the error box, and
clarifies the cause of the error (no webauthn support).

* Replace webauthn with Passkeys