Clean up test-app authentication
This removes the dependency on auth-client for greater flexibility in
tests, removes an unused button (signout) from the test app, and makes
makes the message-display function a bit more robuest/clear.

This will allow us to be more flexible when testing out new APIs, since
we now fully control the authentication flow in the test.