Bump nodejs16 actions
We have a few actions still using the (now deprecated) nodejs16 runtime. This bumps those
actions (setup-node v3 -> v4, cache v3 -> v4) to all use the nodejs20
runtime.

This is unrelated to the node version set up by setup-node, which we use
in our builds. The changes in this PR only impact the nodejs environment
used by the JS-based actions.