Generalize CSP headers (#2248)
This updates the way we create the CSP headers (and in particular, the
`script-src` clause).

Before this, and due to a (now resolved) bug in Firefox, we could not
specify `integrity=` attributes for whitelisting scripts in HTML.
Instead we implemented a work around where we injected a new inline
script (with known hash) which in turn loaded the actual JS script.

This was unconventional, somewhat unsafe (because it didn't actually
include the script's hash) and brittle (we hard coded some script names,
like `index.js`, `index2.js`, etc).

This updates the frontend bundle to include `integrity=` attributes in
HTML script tags, and updates CSP header creation to extract those
integrity hashes and whitelist them in the generated asset list.

The CSP header creation is also changed to keep a reference to the
security headers certified alongside the assets. This ensures the
security headers do not need to be recomputed with every asset, and
clarifies that assets were certified with specific headers.

The CSP header error handling in the test suite is updated for better
error reporting.

Firefox references:
* https://bugzilla.mozilla.org/show_bug.cgi?id=1409200
* https://www.mozilla.org/en-US/firefox/116.0a1/releasenotes/