Don't include dev build checksum in metadata
This changes the dfx metadata to only include a link to the latest
release asset.

Before this, the metadata included the checksum of the dev-flavored
build, which incurred a big complexity cost because the implementation
involved building the dev build before the production build could start.

Since `dfx` now supports not providing a checksum at all, and since the
checksum is mostly useful with big canisters (bigger than II) to prevent
re-downloading a cached canister, we just skip providing a checksum.

The verify-hash script now just performs the build and compare the
hashes with those provided in the arguments.