Move archive buffer to stable memory for freshly initialized II
This implements the archive entries buffer directly in stable memory, but
only if II is initialized from a clean slat.
If the layout v7 is already there, it is maintained and the buffer in
stable memory is just prepared.

A release containing this change can be used as a rollback version after
migration.