Remove redundant memory reserve (#2268)
* Remove redundant memory reserve

When canister memory still was small (and unmanaged) it was sensible
to reserve some space at the end. Now that the canister could in
theory address over 400GB of space _and_ we have more convenient ways
to access that than write into the end adresses, that complexity can
be removed.

The WASM page size constant is also redundant, as it is now provided
by the CDK.

* Add test for default user range