Push newAnchor into LoginData (#2278)
The authentication flow returns whether the user authenticated with a
new or existing anchor. This data is used tailor subsequent flows to new
vs returning users.

Originally this information was added to all return types (success &
error alike) for simplicity, although it only ever makes sense to return
it upon successful authentication (until now we just defaulted to "not a
new anchor" upon failures).

As we start passing more data through the flow (like whether the user
used PIN or not) this shortcut became a bit problematic, so these
changes push the `newAnchor` information deeper into the success state.