Allow authenticating dapps to disable PIN auth (#2298)
* Allow authenticating dapps to disable PIN auth

This extends the II authentication API (postMessage interface) to
include a new boolean field `allowPinAuthentication`. When set to false
(default being true) Internet Identity will not allow the user to
authenticating to the dapp using a PIN identity, and will not allow the
user to register using a PIN identity.

This field is marked as experimental in the spec for now.

* Clarify race