Decouple business logic from stored persistent state (#2326)
* Decouple business logic from stored persistent state

This PR introduces an indirection between the `PersistentState` used in
business logic and the one written to stable memory (in layout version
v9). This allows decoupling refactoring the business logic from storage
layout migrations.

Further, this PR makes sure that all the optionally configurable fields
are initialized and serialized with a sensible default. This allows
removing a lot of `Option` handling in the future.

* Implement review feedback