Switch to dfinity/setup-dfx action
This PR removes our local `setup-dfx` action and replaces it with the one
maintained by the SDK team. This action has the advantage that it integrates
with the new `dfxvm` and reads the desired dfx version from `dfx.json`.

Adding the version there also aligns the dfx version used locally with the
one on CI. The dfx version will be bumped to `0.18.0` in a follow-up
commit.

Side-note on CI runtime: the `dfinity/setup-dfx` does not use caching
(unlike our version) but still takes only about 4s to complete.
Given that we get rid of some complexity that way, I think the 3s are
worth it.

Also, because the `dfinity/setup-dfx` action is not yet properly published
on the market place (there is an open issue for that [here](https://github.com/dfinity/setup-dfx/issues/8))
a specific commit is pinned instead.