Infer issuer canisterId from origin on localhost
For issuers hosted on `localhost` (using the a canister id as a subdomain)
II will no longer make a network request to get the canister ID during
the attribute sharing flow. Instead, the canister id will be parsed from
the origin directly.

This change makes it possible to integrate issuers with
II in a local dev setup if canisterId subdomain
of `localhost` is used as the issuer (derivation) origin.