Refactor add device success screen to be shown consistently
This PR refactors the add device success screen to be shown consistently
(regardless from where the flow is started) by making it part of the
tentative device flow rather than having that logic in the top-level
`index.ts` file.

Given the screen is part of both sides of the flow (the new device _and_
the existing one) I moved the corresponding files out of the `manage`
folder).