Collect anonymous prepare_delegation events for stats purposes (#2449)
* Collect anonymous prepare_delegation events for stats purposes

This PR adds storage to stable memory to keep anonymous(!) events about
`prepare_delegation` calls. This will allow II to create the following
stats:
* number of sign-ins per dapp per ii domain for the last
  * 24h
  * 30d
* cumulative seconds of session time per dapp per ii domain for the last
  * 24h
  * 30d

The latter is necessary to have some comparability between dapps that
use a different session timeout.

The stats are not yet exposed by II. In subsequent PRs:
1. the old collection of observed front-end origins will be removed
2. the new data will be exposed
3. integration tests will be added
4. additional performance metrics will be collected in order to better
   judge the impact it has on the `prepare_delegation` call.

* Remove aggregations with zero weight on update

* Implement review feedback

* Change delimiter to character that is forbidden in URLs

Switches aggregation key section delimiter to `>` (from `_`).

* Ensure unique key for events with same timestamp

This commit introduces a `EventKey` struct to use as a key for `event_data`
as the raw timestamp does not provide uniqueness when to events are added
in the same execution round.

* Represent aggregation as trait and use structured aggregation key

This makes the code easier to follow by representing aggregations as
trait objects and using properly typed aggregation keys.

* Clippy fixes