Implement automatic event pruning (#2460)
This PR adds automatic event pruning using timers and self-calls (to
enable retries on failure).

In normal operation automatic event pruning should not be needed.
However, if II stops processing `prepare_delegation` canister calls for
some reason for an extended period of time then the backlog of events
that need pruning will grow.

To make sure that II can process this backlog, `PruneEvents` are injected
if the latest event is older than 5 minutes. If the backlog is too large
to be processed in a single canister call, then II will fall back to
smaller time windows until it becomes processable.