Add explicit bound to amortized clean-up (#2489)
* Add explicit bound to amortized clean-up

This adds an explicit limit to the number of events processed during
amortized clean-up. Using that, we can (in a future PR) remove the timer
based mechanism for pruning data.

* Remove notion of pruning for 24h event aggregations window

Since the underlying event data is not deleted, the word "prune" is
misleading in the context. It is replaced with "remove from aggregation".

* Clarify function name and comments