Add support to provision dummy issuer with wallet (#2500)
In #2493 a guard was introduced on the dummy issuer to only allow
configure to be called by controllers.
Because the current mainnet deployment of the dummy issuer is only
controlled by the wallet `cvthj-wyaaa-aaaad-aaaaq-cai`, this PR extends
the provision script to optionally allow making the configure call
through the wallet.