Add support to reconfigure derivation_origin for VC issuer
This PR allows to reconfigure the expected frontend hostname as well as
the derivation origin. This will be used in the future to simplify the
e2e tests.

Since this is most useful in e2e tests if the configuration is _not_
guarded by the controller, the guard is removed for the `configure`
method too for consistency.