Refactor VC flow e2e tests (#2532)
* Refactor VC flow e2e tests

This PR refactors the VC related e2e tests such that the test_app is no
longer misused to test alternative origins configurations for the issuer.

And because the issuer can now be configured as needed in tests, the call to
the provision script in `dfx.json` no longer needs to configure the
test app as the issuer derivation origin either.

Additionally a new test is added to verify that the VC flow indeed fails
on issuer alternative origins misconfiguration.

* Extract function to create issuer actor and remove unused principal registration

* Add issuer reset to default VC flow tests