Fix ic-wasm action failing on wrong version
If `ic-wasm` is not detected properly (this was encountered when caching
actions became unreliable due to hitting the rate limit), then the
bootstrap script attempts to install the right version. However, if cargo
is not aware that a binary already exists, it will then fail.
By using `--force` we allow the script to continue an override the
existing binary.