Refactor temp_keys in preparation for dynamic captcha registration flow
For the dynamic captcha registration flow, temp_keys need to be tracked
before the identity number or authn method public key is known. For that
reason, this refactoring abstracts out the `TempKeyId` such that we can
introduce a different variant to track the keys.
For example a registration flow id.