Track registration rate and expose metrics (#2634)
* Track registration rate and expose metrics

This PR introduces the functionality of tracking the current and the
reference rate of new registrations.

In addition, it will also always calculate and expose the threshold rate
that would need to be crossed to trigger the captcha being activated.
This will be useful to later analyze how often the captcha was actually
shown.

Note: This is only the tracking / metrics. The captcha itslef is not yet
dynamic.

* Improved comments based on review feedback

* Doc indentation

* Cargo fmt