Make identity metadata update synchronous
There is no need to `await` metadata if some value needs to be written.
Instead we just update the internal promise of the metadata repository
to update the value once it is loaded.

This also makes a change to the registration flow to make sure that when
the success page is shown that the metadata is committed to make sure
the identity is fully set up. This should finally resolve the observed
flakiness in the e2e tests.