Switch II front-end to API v2 for registration (#2654)
This PR switches the II front-end to the APIv2. In particular, the
front-end is now able to dynamically skip the captcha if it is switched
off.

Note: This PR does _not_ include e2e tests for switched off captchas.
Since this requires a whole new CI setup, I'll do this in another PR.