fix: Special device migration (#3530)
# Motivation

This PR adds special case handling to the anchor device migration. 

**Context**

With the previous migration code in `impl From<Anchor> for
(StorableFixedAnchor, StorableAnchor)`, all anchors were migrated to
unbounded memory (which is not currently used). During this process,
some inconsistent keys were discovered.
* credential_id = null (60135 cases)
    * key type = BrowserStorageKey (56822 cases)
    * key type = Unknown (3308 cases)
    * key type = Platform (5 cases)

This PR adds a handler for all the special cases above to the migration
and repeat the migration in the next release.

Main condition: if there is a credential ID, fall back to passkey;
otherwise, treat this as a recovery phrase.

For the special cases, the (credential_id, purpose, key_type) triplets
is associated with the migrated data, so there is no risk of data loss.

# Changes

* Introduced `SpecialDeviceMigration` struct to track inconsistent
device configurations during migration.
* Added optional `special_device_migration` field to both
`StorablePasskeyCredential` and `StorableRecoveryKey` structs.
* Updated migration logic in `From<Anchor>` implementation to handle
edge cases explicitly.

# Tests

* Adjusted to cover some inconsistent corner cases
(`should_handle_devices_without_credential_id`).
<!-- SCREENSHOTS REPORT START -->

<!-- SCREENSHOTS REPORT STOP -->